# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Estate Hub — Real Estate ERP + CRM SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: getCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a single customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PROPERTIES ─────────────────────────────────────────────
...
  /installments:
    get:
      operationId: getInstallments
      tags: [installments]
      summary: List installments
      parameters:
        - name: payment_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: customer_id
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of installments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Installment"
    post:
      operationId: createInstallment
      tags: [installments]
      summary: Create an installment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallmentInput"
      responses:
        "201":
          description: Created installment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installment"
  /installments/{id}:
    get:
      operationId: getInstallment
      tags: [installments]
      summary: Get a single installment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Installment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installment"
    patch:
      operationId: updateInstallment
      tags: [installments]
      summary: Update an installment (e.g. mark paid)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallmentUpdate"
      responses:
        "200":
          description: Updated installment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installment"
  /installments/upcoming:
    get:
      operationId: getUpcomingInstallments
      tags: [installments]
      summary: Get upcoming installments (due within 30 days)
      parameters:
        - name: customer_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Upcoming installments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Installment"

  # ─── USERS ──────────────────────────────────────────────────
...
  /documents:
    get:
      operationId: getDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: customer_id
          in: query
          schema:
            type: integer
        - name: property_id
          in: query
          schema:
            type: integer
        - name: doc_type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── AUDIT LOGS ──────────────────────────────────────────────
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        company_id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        nid:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - company_id
        - name
        - phone
        - created_at
        - updated_at
    CustomerInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        nid:
          type: string
        notes:
          type: string
      required:
        - name
        - phone
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        nid:
          type: string
        notes:
          type: string
    CustomerListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit

    # ─── PROPERTY SCHEMAS ────────────────────────────────────
...
    Installment:
      type: object
      properties:
        id:
          type: integer
        payment_id:
          type: integer
        company_id:
          type: integer
        installment_no:
          type: integer
        amount:
          type: number
        due_date:
          type: string
        paid_date:
          type: ["string", "null"]
        status:
          type: string
        customer_name:
          type: ["string", "null"]
        property_name:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - payment_id
        - company_id
        - installment_no
        - amount
        - due_date
        - status
        - created_at
        - updated_at
    InstallmentInput:
      type: object
      properties:
        payment_id:
          type: integer
        installment_no:
          type: integer
        amount:
          type: number
        due_date:
          type: string
        status:
          type: string
      required:
        - payment_id
        - installment_no
        - amount
        - due_date
    InstallmentUpdate:
      type: object
      properties:
        amount:
          type: number
        due_date:
          type: string
        paid_date:
          type: string
        status:
          type: string

    # ─── USER SCHEMAS ────────────────────────────────────────
...
    Document:
      type: object
      properties:
        id:
          type: integer
        company_id:
          type: integer
        customer_id:
          type: ["integer", "null"]
        property_id:
          type: ["integer", "null"]
        customer_name:
          type: ["string", "null"]
        property_name:
          type: ["string", "null"]
        doc_type:
          type: string
        file_name:
          type: string
        file_url:
          type: string
        uploaded_by:
          type: ["integer", "null"]
        created_at:
          type: string
      required:
        - id
        - company_id
        - doc_type
        - file_name
        - file_url
        - created_at
    DocumentInput:
      type: object
      properties:
        customer_id:
          type: ["integer", "null"]
        property_id:
          type: ["integer", "null"]
        doc_type:
          type: string
        file_name:
          type: string
        file_url:
          type: string
      required:
        - doc_type
        - file_name
        - file_url

    # ─── AUDIT LOG SCHEMAS ───────────────────────────────────
...